#21 수업관리 및 채팅방 복구
