+++ src/router/Router.js
@@ -18,7 +18,6 @@ import CreateLecture from "../pages/CreateLecture";
 import GetLecture from "../pages/GetLecture";
 
 import Lectures from "../pages/Lectures";
-import ManageLecture from "../pages/ManageLecture";
 import MyPage from "../pages/MyPage";
 import FindTutor from "../routes/findtutor/FindTutor";
 
@@ -38,8 +37,6 @@ const Router = () => {
                 <Route path="/lectures/:id" element={<GetLecture />} />
 
                 <Route path="/mypage" element={<MyPage />} />
-
-                <Route path="/manage/lecture" element={<ManageLecture />} />
 
                 <Route path="/chat" element={<ChatRoom />} />
